Rename sunSafePoint modules for clarify
