+++ src/fswAlgorithms/attGuidance/sunSafePoint/sunSafePoint.cpp
@@ -1,7 +1,7 @@
 /*
  ISC License
 
- Copyright (c) 2016, Autonomous Vehicle Systems Lab, University of Colorado at Boulder
+ Copyright (c) 2024, Laboratory for Atmospheric and Space Physics, University of Colorado at Boulder
 
  Permission to use, copy, modify, and/or distribute this software for any
  purpose with or without fee is hereby granted, provided that the above
@@ -17,119 +17,196 @@
 
  */
 
+#include "sunSafePoint.h"
+
 #include <math.h>
-#include "fswAlgorithms/attGuidance/sunSafePoint/sunSafePoint.h"
+
+#include "architecture/utilities/avsEigenSupport.h"
 #include "architecture/utilities/linearAlgebra.h"
 #include "architecture/utilities/rigidBodyKinematics.h"
 
 /*! This method performs a complete reset of the module.  Local module variables that retain
  time varying states between function calls are reset to their default values.
  @return void
- @param callTime The clock time at which the function was called (nanoseconds)
- */
-void SunSafePoint::reset(uint64_t callTime)
-{
-    double v1[3];
-
-    // check if the required input messages are included
+ @param callTime [ns] Time the method is called
+*/
+void SunSafePoint::reset(uint64_t callTime) {
+    // Check if the required input messages are linked
     if (!this->sunDirectionInMsg.isLinked()) {
-        this->bskLogger.bskLog(BSK_ERROR, "Error: sunSafePoint.sunDirectionInMsg wasn't connected.");
+        _bskLog(this->bskLogger, BSK_ERROR, "sunSafePoint.sunDirectionInMsg wasn't connected.");
     }
     if (!this->imuInMsg.isLinked()) {
-        this->bskLogger.bskLog(BSK_ERROR, "Error: sunSafePoint.imuInMsg wasn't connected.");
+        _bskLog(this->bskLogger, BSK_ERROR, "sunSafePoint.imuInMsg wasn't connected.");
     }
 
-    /* compute an Eigen axis orthogonal to sHatBdyCmd */
-    if (v3Norm(this->sHatBdyCmd)  < 0.1) {
-      char info[MAX_LOGGING_LENGTH];
-      snprintf(info, sizeof(info), "The module vector sHatBdyCmd is not setup as a unit vector [%f, %f %f]",
-                this->sHatBdyCmd[0], this->sHatBdyCmd[1], this->sHatBdyCmd[2]);
-      this->bskLogger.bskLog(BSK_ERROR, info);
+    // Compute an Eigen axis orthogonal to sHatBdyCmd
+    if (this->sHatBdyCmd.norm() < 0.1) {
+        char info[MAX_LOGGING_LENGTH];
+        snprintf(info,
+                 sizeof(info),
+                 "The module vector sHatBdyCmd is not setup as a unit vector [%f, %f %f]",
+                 this->sHatBdyCmd[0],
+                 this->sHatBdyCmd[1],
+                 this->sHatBdyCmd[2]);
+        _bskLog(this->bskLogger, BSK_ERROR, info);
     } else {
-        v3Set(1., 0., 0., v1);
-        v3Normalize(this->sHatBdyCmd, this->sHatBdyCmd);    /* ensure that this vector is a unit vector */
-        v3Cross(this->sHatBdyCmd, v1, this->eHat180_B);
-        if (v3Norm(this->eHat180_B) < 0.1) {
-            v3Set(0., 1., 0., v1);
-            v3Cross(this->sHatBdyCmd, v1, this->eHat180_B);
+        Eigen::Vector3d v1 = {1.0, 0.0, 0.0};
+        this->sHatBdyCmd = this->sHatBdyCmd / this->sHatBdyCmd.norm();  // Ensure that this vector is a unit vector
+        this->eHat180_B = this->sHatBdyCmd.cross(v1);
+        if (this->eHat180_B.norm() < 0.1) {
+            v1 = {0.0, 1.0, 0.0};
+            this->eHat180_B = this->sHatBdyCmd.cross(v1);
         }
-        v3Normalize(this->eHat180_B, this->eHat180_B);
+        this->eHat180_B = this->eHat180_B / this->eHat180_B.norm();
     }
-
-    this->attGuidanceOutBuffer = {};
-
-    return;
 }
 
 /*! This method takes the estimated body-observed sun vector and computes the
  current attitude/attitude rate errors to pass on to control.
  @return void
- @param callTime The clock time at which the function was called (nanoseconds)
+ @param callTime [ns] Time the method is called
 */
-void SunSafePoint::updateState(uint64_t callTime)
-{
-    NavAttMsgPayload navMsg;
-    double ctSNormalized;
-    double sNorm;                   /*!< --- Norm of measured direction vector */
-    double e_hat[3];                /*!< --- Eigen Axis */
-    double omega_BN_B[3];           /*!< r/s inertial body angular velocity vector in B frame components */
-    double omega_RN_B[3];           /*!< r/s local copy of the desired reference frame rate */
+void SunSafePoint::updateState(uint64_t callTime) {
+    // Zero the attitude guidance output buffer message
+    this->attGuidanceOutBuffer = AttGuidMsgPayload();
 
-    NavAttMsgPayload localImuDataInBuffer;
-    this->attGuidanceOutBuffer = {};
-
-    /*! - Read the current sun body vector estimate*/
-    navMsg = this->sunDirectionInMsg();
-    localImuDataInBuffer = this->imuInMsg();
-
-    v3Copy(localImuDataInBuffer.omega_BN_B, omega_BN_B);
-
-    /*! - Compute the current error vector if it is valid*/
-    sNorm = v3Norm(navMsg.vehSunPntBdy);
-    if(sNorm > this->minUnitMag)
-    {
-        /* a good sun direction vector is available */
-        ctSNormalized = v3Dot(this->sHatBdyCmd, navMsg.vehSunPntBdy)/sNorm;
-        ctSNormalized = fabs(ctSNormalized) > 1.0 ?
-        ctSNormalized/fabs(ctSNormalized) : ctSNormalized;
-        this->sunAngleErr = safeAcos(ctSNormalized);
-
-        /*
-            Compute the heading error relative to the sun direction vector
-         */
-        if (this->sunAngleErr < this->smallAngle) {
-            /* sun heading and desired body axis are essentially aligned.  Set attitude error to zero. */
-             v3SetZero(this->attGuidanceOutBuffer.sigma_BR);
-        } else {
-            if (M_PI - this->sunAngleErr < this->smallAngle) {
-                /* the commanded body vector nearly is opposite the sun heading */
-                v3Copy(this->eHat180_B, e_hat);
-            } else {
-                /* normal case where sun and commanded body vectors are not aligned */
-                v3Cross(navMsg.vehSunPntBdy, this->sHatBdyCmd, e_hat);
-            }
-            v3Normalize(e_hat, this->sunMnvrVec);
-            v3Scale(tan(this->sunAngleErr*0.25), this->sunMnvrVec,
-                    this->attGuidanceOutBuffer.sigma_BR);
-            MRPswitch(this->attGuidanceOutBuffer.sigma_BR, 1.0, this->attGuidanceOutBuffer.sigma_BR);
-        }
-
-        /* rate tracking error are the body rates to bring spacecraft to rest */
-        v3Scale(this->sunAxisSpinRate/sNorm, navMsg.vehSunPntBdy, omega_RN_B);
-        v3Subtract(omega_BN_B, omega_RN_B, this->attGuidanceOutBuffer.omega_BR_B);
-        v3Copy(omega_RN_B, this->attGuidanceOutBuffer.omega_RN_B);
-
-    } else {
-        /* no proper sun direction vector is available */
-        v3SetZero(this->attGuidanceOutBuffer.sigma_BR);
-
-        /* specify a body-fixed constant search rotation rate */
-        v3Subtract(omega_BN_B, this->omega_RN_B, this->attGuidanceOutBuffer.omega_BR_B);
-        v3Copy(this->omega_RN_B, this->attGuidanceOutBuffer.omega_RN_B);
+    // Read the current sun body vector estimate input message
+    this->sunDirectionInBuffer = NavAttMsgPayload();
+    if (this->sunDirectionInMsg.isWritten()) {
+        this->sunDirectionInBuffer = this->sunDirectionInMsg();
     }
 
-    /* write the Guidance output message */
-    this->attGuidanceOutMsg.write(&this->attGuidanceOutBuffer, this->moduleID, callTime);
+    // Determine norm of measured Sun-direction vector
+    double sHatNorm = cArray2EigenVector3d(this->sunDirectionInBuffer.vehSunPntBdy).norm();
 
-    return;
+    // Computing the attitude guidance states sigma_BR and omega_RN_B
+    if (this->sunDirectionIsAvailable(sHatNorm)) {
+        this->computeAttGuidanceStates(sHatNorm);
+    } else {
+        Eigen::Vector3d sigma_BR = Eigen::Vector3d::Zero();
+        eigenVector3d2CArray(sigma_BR, this->attGuidanceOutBuffer.sigma_BR);
+    }
+
+    // Compute the hub angular rate error omega_BR_B
+    this->computeHubAngularRateError();
+
+    // Write the guidance output message
+    eigenVector3d2CArray(this->omega_RN_B, this->attGuidanceOutBuffer.omega_RN_B);
+    this->attGuidanceOutMsg.write(&this->attGuidanceOutBuffer, moduleID, callTime);
 }
+
+/*! Method for computing the attitude guidance states sigma_BR and omega_RN_B if a valid sun direction vector is
+ available.
+ @return void
+ @param sHatNorm Norm of measured Sun-direction vector
+*/
+void SunSafePoint::computeAttGuidanceStates(double sHatNorm) {
+    double dotProductNormalized =
+        this->sHatBdyCmd.dot(cArray2EigenVector3d(this->sunDirectionInBuffer.vehSunPntBdy)) / sHatNorm;
+    dotProductNormalized =
+        fabs(dotProductNormalized) > 1.0 ? dotProductNormalized / fabs(dotProductNormalized) : dotProductNormalized;
+    this->sunAngleErr = safeAcos(dotProductNormalized);
+
+    // Compute the heading error relative to the sun direction vector
+    if (this->sunAngleErr <
+        this->smallAngle) {  // Sun heading and desired body axis are essentially aligned. Set attitude error to zero.
+        Eigen::Vector3d sigma_BR = Eigen::Vector3d::Zero();
+        eigenVector3d2CArray(sigma_BR, this->attGuidanceOutBuffer.sigma_BR);
+    } else {
+        Eigen::Vector3d e_hat;  // Eigen Axis
+        if (M_PI - this->sunAngleErr <
+            this->smallAngle) {  // The commanded body vector nearly is opposite the sun heading
+            e_hat = this->eHat180_B;
+        } else {  // Normal case where sun and commanded body vectors are not aligned
+            e_hat = cArray2EigenVector3d(this->sunDirectionInBuffer.vehSunPntBdy).cross(this->sHatBdyCmd);
+        }
+        this->sunMnvrVec = e_hat / e_hat.norm();
+        Eigen::Vector3d v2 = tan(this->sunAngleErr * 0.25) * this->sunMnvrVec;
+        eigenVector3d2CArray(v2, this->attGuidanceOutBuffer.sigma_BR);
+        MRPswitch(this->attGuidanceOutBuffer.sigma_BR, 1.0, this->attGuidanceOutBuffer.sigma_BR);
+    }
+
+    // Rate tracking error is the body rate to bring spacecraft to rest
+    this->omega_RN_B =
+        (this->sunAxisSpinRate / sHatNorm) * cArray2EigenVector3d(this->sunDirectionInBuffer.vehSunPntBdy);
+}
+
+/*! Method for computing the hub angular rate error omega_BR_B.
+ @return void
+*/
+void SunSafePoint::computeHubAngularRateError() {
+    // Read the imu guidance input message
+    NavAttMsgPayload localImuDataInBuffer = NavAttMsgPayload();
+    if (this->imuInMsg.isWritten()) {
+        localImuDataInBuffer = this->imuInMsg();
+    }
+
+    // Create local copy of hub inertial angular velocity vector in B frame components
+    Eigen::Vector3d omega_BN_B = cArray2EigenVector3d(localImuDataInBuffer.omega_BN_B);  // [rad/s]
+
+    // Create local copy of hub angular velocity error in B frame components
+    Eigen::Vector3d omega_BR_B = omega_BN_B - this->omega_RN_B;  // [rad/s]
+
+    eigenVector3d2CArray(omega_BR_B, this->attGuidanceOutBuffer.omega_BR_B);
+}
+
+/*! Method for determining if a valid sun direction vector is available.
+ @return bool
+ @param sHatNorm Norm of measured Sun-direction vector
+*/
+bool SunSafePoint::sunDirectionIsAvailable(double sHatNorm) const { return sHatNorm > this->minUnitMag; }
+
+/*! Getter method for the minimally accepted sun body vector norm.
+ @return double
+*/
+double SunSafePoint::getMinUnitMag() const { return this->minUnitMag; }
+
+/*! Getter method for the small alignment tolerance angle near 0 or 180 degrees.
+ @return double
+*/
+double SunSafePoint::getSmallAngle() const { return this->smallAngle; }
+
+/*! Getter method for the desired constant spin rate about sun heading vector.
+ @return double
+*/
+double SunSafePoint::getSunAxisSpinRate() const { return this->sunAxisSpinRate; }
+
+/*! Getter method for the desired body rate vector if no sun direction is available.
+ @return const Eigen::Vector3d
+*/
+const Eigen::Vector3d &SunSafePoint::getOmega_RN_B() const { return this->omega_RN_B; }
+
+/*! Getter method for the desired body vector to point at the sun.
+ @return const Eigen::Vector3d
+*/
+const Eigen::Vector3d &SunSafePoint::getSHatBdyCmd() const { return this->sHatBdyCmd; }
+
+/*! Setter method for the minimally accepted sun body vector norm.
+ @return void
+ @param minUnitMag The minimally acceptable norm of sun body vector
+*/
+void SunSafePoint::setMinUnitMag(const double minUnitMag) { this->minUnitMag = minUnitMag; }
+
+/*! Setter method for the small alignment tolerance angle near 0 or 180 degrees.
+ @return void
+ @param smallAngle [rad] An angle value that specifies what is near 0 or 180 degrees
+*/
+void SunSafePoint::setSmallAngle(const double smallAngle) { this->smallAngle = smallAngle; }
+
+/*! Setter method for the desired constant spin rate about sun heading vector.
+ @return void
+ @param sunAxisSpinRate [rad/s] Desired constant spin rate about sun heading vector
+*/
+void SunSafePoint::setSunAxisSpinRate(const double sunAxisSpinRate) { this->sunAxisSpinRate = sunAxisSpinRate; }
+
+/*! Setter method for the desired body rate vector if no sun direction is available.
+ @return void
+ @param omega_RN_B [rad/s] Desired body rate vector if no sun direction is available
+*/
+void SunSafePoint::setOmega_RN_B(const Eigen::Vector3d &omega_RN_B) { this->omega_RN_B = omega_RN_B; }
+
+/*! Setter method for the desired body vector to point at the sun.
+ @return void
+ @param sHatBdyCmd Desired body vector to point at the sun
+*/
+void SunSafePoint::setSHatBdyCmd(const Eigen::Vector3d &sHatBdyCmd) { this->sHatBdyCmd = sHatBdyCmd; }
