+++ src/fswAlgorithms/attGuidance/sunSafePoint/sunSafePoint.h
@@ -1,7 +1,7 @@
 /*
  ISC License
 
- Copyright (c) 2016, Autonomous Vehicle Systems Lab, University of Colorado at Boulder
+ Copyright (c) 2024, Laboratory for Atmospheric and Space Physics, University of Colorado at Boulder
 
  Permission to use, copy, modify, and/or distribute this software for any
  purpose with or without fee is hereby granted, provided that the above
@@ -22,33 +22,62 @@
 
 #include "architecture/_GeneralModuleFiles/sys_model.h"
 #include "architecture/messaging/messaging.h"
-#include "architecture/msgPayloadDefC/NavAttMsgPayload.h"
 #include "architecture/msgPayloadDefC/AttGuidMsgPayload.h"
-
+#include "architecture/msgPayloadDefC/NavAttMsgPayload.h"
+#include "architecture/utilities/avsEigenSupport.h"
 #include "architecture/utilities/bskLogging.h"
 #include <stdint.h>
+#include <Eigen/Dense>
 
-
-
-/*! @brief Top level structure for the sun-safe attitude guidance routine.*/
+/*! @brief Sun safe point attitude guidance class. */
 class SunSafePoint : public SysModel {
-public:
-    void reset(uint64_t callTime) override;
-    void updateState(uint64_t callTime) override;
-    Message<AttGuidMsgPayload> attGuidanceOutMsg; /*!< The name of the output message*/
-    ReadFunctor<NavAttMsgPayload> sunDirectionInMsg; /*!< The name of the Input message*/
-    ReadFunctor<NavAttMsgPayload> imuInMsg;    /*!< The name of the incoming IMU message*/
-    double minUnitMag;       /*!< -- The minimally acceptable norm of sun body vector*/
-    double sunAngleErr;      /*!< rad The current error between cmd and obs sun angle*/
-    double smallAngle;       /*!< rad An angle value that specifies what is near 0 or 180 degrees */
-    double eHat180_B[3];     /*!< -- Eigen axis to use if commanded axis is 180 from sun axis */
-    double sunMnvrVec[3];    /*!< -- The eigen axis that we want to rotate on to get sun*/
-    double sHatBdyCmd[3];    /*!< -- Desired body vector to point at the sun*/
-    double omega_RN_B[3];    /*!< -- Desired body rate vector if no sun direction is available */
-    double sunAxisSpinRate;  /*!< r/s Desired constant spin rate about sun heading vector */
+   public:
+    SunSafePoint() = default;   //!< Constructor
+    ~SunSafePoint() = default;  //!< Destructor
 
-    AttGuidMsgPayload attGuidanceOutBuffer;   /*!< -- The output data that we compute*/
-    BSKLogger bskLogger={};                             //!< BSK Logging
+    void reset(uint64_t currentSimNanos) override;        //!< Reset member function
+    void updateState(uint64_t currentSimNanos) override;  //!< Update member function
+
+    double getMinUnitMag() const;       //!< Getter method for the minimally accepted sun body vector norm
+    double getSmallAngle() const;       //!< Getter method for the small alignment tolerance angle near 0 or 180 degrees
+    double getSunAxisSpinRate() const;  //!< Getter method for the desired constant spin rate about sun heading vector
+    const Eigen::Vector3d &getOmega_RN_B()
+        const;  //!< Getter method for the desired body rate vector if no sun direction is available
+    const Eigen::Vector3d &getSHatBdyCmd() const;  //!< Getter method for the desired body vector to point at the sun
+    void setMinUnitMag(const double minUnitMag);   //!< Setter method for the minimally accepted sun body vector norm
+    void setSmallAngle(
+        const double smallAngle);  //!< Setter method for the small alignment tolerance angle near 0 or 180 degrees
+    void setSunAxisSpinRate(
+        const double sunAxisSpinRate);  //!< Setter method for the desired constant spin rate about sun heading vector
+    void setOmega_RN_B(const Eigen::Vector3d &omega_RN_B);  //!< Setter method for the desired body rate vector if no
+                                                            //!< sun direction is available
+    void setSHatBdyCmd(
+        const Eigen::Vector3d &sHatBdyCmd);  //!< Setter method for the desired body vector to point at the sun
+
+    ReadFunctor<NavAttMsgPayload> imuInMsg;           //!< IMU attitude guidance input message
+    ReadFunctor<NavAttMsgPayload> sunDirectionInMsg;  //!< Sun attitude guidance input message
+    Message<AttGuidMsgPayload> attGuidanceOutMsg;     //!< Attitude guidance output message
+
+    BSKLogger *bskLogger;  //!< BSK Logging
+
+   private:
+    void computeAttGuidanceStates(double sHatNorm);  //!< Method for computing the attitude guidance states sigma_BR and
+                                                     //!< omega_RN_B if a valid sun direction vector is available
+    void computeHubAngularRateError();               //!< Method for computing the hub angular rate error omega_BR_B
+    bool sunDirectionIsAvailable(
+        double sHatNorm) const;  //!< Method for determining if a valid sun direction vector is available
+
+    double minUnitMag;           //!< The minimally acceptable norm of sun body vector
+    double smallAngle;           //!< [rad] An angle value that specifies what is near 0 or 180 degrees
+    double sunAxisSpinRate;      //!< [rad/s] Desired constant spin rate about sun heading vector
+    double sunAngleErr;          //!< [rad] The current error between cmd and obs sun angle
+    Eigen::Vector3d omega_RN_B;  //!< [rad/s] Desired body rate vector if no sun direction is available
+    Eigen::Vector3d sHatBdyCmd;  //!< Desired body vector to point at the sun
+    Eigen::Vector3d eHat180_B;   //!< Eigen axis to use if commanded axis is 180 from sun axis
+    Eigen::Vector3d sunMnvrVec;  //!< The Eigen axis that we want to rotate on to get sun
+
+    AttGuidMsgPayload attGuidanceOutBuffer;  //!< Attitude guidance output message buffer
+    NavAttMsgPayload sunDirectionInBuffer;   //!< Sun attitude guidance input message buffer
 };
 
 #endif
