+++ src/fswAlgorithms/attGuidance/sunSafePoint/sunSafePoint.i
@@ -22,16 +22,19 @@
 %}
 
 %pythoncode %{
-    from Basilisk.architecture.swig_common_model import *
+from Basilisk.architecture.swig_common_model import *
 %}
+%include "std_string.i"
+%include "swig_conly_data.i"
+%include "swig_eigen.i"
 
 %include "sys_model.i"
-%include "swig_conly_data.i"
-
 %include "sunSafePoint.h"
 
 %include "architecture/msgPayloadDefC/NavAttMsgPayload.h"
+struct NavAttMsg_C;
 %include "architecture/msgPayloadDefC/AttGuidMsgPayload.h"
+struct AttGuidMsg_C;
 
 %pythoncode %{
 import sys
