--- src/fswAlgorithms/attGuidance/sunSafePointCpp/_Documentation/secModelFunctions.tex
@@ -1,15 +1,0 @@
-% !TEX root = ./Basilisk-sunSafePoint-20180427.tex
-
-
-\section{Module Functions}
-\begin{itemize}
-	\item \textbf{Compute the attitude tracking error}: Determines the shortest rotation to align $\bm s$ and $\hat{\bm s}_{c}$, and computes the corresponding three-dimensional attitude difference
-	\item \textbf{Control spacecraft rotation}: The reference frame is assumed to be non-accelerating and either zero (default) or a constant spin about the sun heading axis.
-	
-	\item \textbf{Robust to no sun heading information}: If the sun heading vector is not available, then the attitude feedback is turned off by zeroing $\bm\sigma_{B/R}$.  Instead of driving the body rates to zero, the body rates are driven to a prescribed $\bm\omega_{R/N}$ vector.  This allow the spacecraft to search for the sun and covers the case if some sun sensors are offline.  
-	
-	\item \textbf{Robust to collinear observation vector}:  The module must handle the cases where the commanded body relative vector and the sun direction vectors are nearly collinear. 
-\end{itemize}
-
-\section{Module Assumptions and Limitations}
-The module input vector $\bm s$ can be a vector of any length except for a zero-length vector.  The commanded body-relative unit direction vector ${\bm s}_{c}$ is assumed to be fixed relative to the body.
