--- src/fswAlgorithms/attGuidance/sunSafePointCpp/_Documentation/secTest.tex
@@ -1,66 +1,0 @@
-% !TEX root = ./Basilisk-sunSafePoint-20180427.tex
-
-\section{Test Description and Success Criteria}
-The mathematics in this module are straight forward and can be tested in a series of input and output evaluation tests.
-
-
-\subsection{Check 1}
-Here a check is performed where the sun vector measurement $\bm s$ has a non-zero length and is not aligned with $\hat{\bm s}_{c}$.  
-
-\subsection{Check 2}
-The sun direction vector $\bm s$ is given a norm value that is less than {\tt minUnitMag}.  In this case the attitude tracking $\bm\sigma_{B/R}$ should be set to zero.  Further, the body rate errors are now evaluated relative to a fixed $\bm\omega_{R/N}$ vector.  
-
-\subsection{Check 3}
-The sun direction vector $\bm s$ aligned with $\hat{\bm s}_{c}$.  In this case the attitude tracking $\bm\sigma_{B/R}$ should be set to zero.  Further, the body rate errors are simply the inertial body angular rates.  
-
-\subsection{Check 4}
-The sun direction vector $\bm s \approx -\hat{\bm s}_{c}$.  In this case the attitude tracking $\bm\sigma_{B/R}$ should be set to $\hat{\bm e}_{180}$.  Further, the body rate errors are simply the inertial body angular rates.  
-
-\subsection{Check 4}
-The sun direction vector $\bm s \approx -\hat{\bm s}_{c}$, but $\hat{\bm s}_{c} = \hat{\bm b}_{1}$.  In this case the attitude tracking $\bm\sigma_{B/R}$ should be set to $\hat{\bm e}_{180}$ that is evaluated with the cross product with $\hat{\bm b}_{2}$.  Further, the body rate errors are simply the inertial body angular rates.  
-
-
-\section{Test Parameters}
-The unit test verify that the module output guidance message vectors match expected values.
-\begin{table}[htbp]
-	\caption{Error tolerance for each test.}
-	\label{tab:errortol}
-	\centering \fontsize{10}{10}\selectfont
-	\begin{tabular}{ c | c } % Column formatting, 
-		\hline\hline
-		\textbf{Output Value Tested}  & \textbf{Tolerated Error}  \\ 
-		\hline
-		$\bm\sigma_{B/R}$        & \input{AutoTex/toleranceValue}	   \\ 
-		$\bm\omega_{B/R}$        & \input{AutoTex/toleranceValue}\\ 
-		$\bm\omega_{R/N}$        & \input{AutoTex/toleranceValue} \\ 
-		$\dot{\bm\omega}_{R/N}$        & \input{AutoTex/toleranceValue}  \\ 
-		\hline\hline
-	\end{tabular}
-\end{table}
-
-The nominal module test input values are $\hat{\bm s}_{c} = (0,0,1)$, $\bm s = (1,0,0)$ and $\leftexp{B}{\bm\omega}_{B/N} = (0.01, 0.50, -0.20)$ rad/sec.  The nominal body-fixed search rate is set to $\leftexp{B}{\bm\omega}_{R/N} = (0.0, 0.0, 0.1)$ rad/sec.  This rate is only used if no sun direction vector is available.  The small angle is set to $\epsilon$ = 0.01 degrees.  
-
-
-
-\section{Test Results}
-
-All of the tests passed:
-\begin{table}[H]
-	\caption{Test results}
-	\label{tab:results}
-	\centering \fontsize{10}{10}\selectfont
-	\begin{tabular}{c | c  } % Column formatting, 
-		\hline\hline
-		\textbf{Check} 						  		&\textbf{Pass/Fail} \\ 
-		\hline
-	   1	   			& \input{AutoTex/passFail1} \\ 
-	   2	   			& \input{AutoTex/passFail2} \\ 
-	   3	   			& \input{AutoTex/passFail3} \\ 
-	   4	   			& \input{AutoTex/passFail4} \\ 
-	   5	   			& \input{AutoTex/passFail5} \\ 
-	   \hline\hline
-	\end{tabular}
-\end{table}
-
-
-
