+++ src/fswAlgorithms/attGuidance/sunSafePoint_C/_Documentation/AVS.sty
@@ -1,6 +1,6 @@
 %
 %	AVS.sty
-%	
+%
 %	provides common packages used to edit LaTeX papers within the AVS lab
 %  	and creates some common mathematical macros
 %
@@ -19,7 +19,7 @@
 
 %
 %	define Track changes macros and colors
-% 
+%
 \definecolor{colorHPS}{rgb}{1,0,0} % Red
 \definecolor{COLORHPS}{rgb}{1,0,0} % Red
 \definecolor{colorPA}{rgb}{1,0,1} % Bright purple
@@ -94,5 +94,3 @@
 %	define the oe orbit element symbol for the TeX math environment
 \renewcommand{\OE}{{o\hspace*{-2pt}e}}
 \renewcommand{\oe}{{\bm o\hspace*{-2pt}\bm e}}
-
-
