+++ src/fswAlgorithms/attGuidance/sunSafePoint_C/_Documentation/Basilisk-sunSafePoint-20180427.pdf
