+++ src/fswAlgorithms/attGuidance/sunSafePoint_C/_Documentation/Basilisk-sunSafePoint-20180427.tex
@@ -93,7 +93,7 @@
 
 
 
-	
+
 \input{secModelDescription.tex} %This section includes mathematical models, code description, etc.
 
 \input{secModelFunctions.tex} %This includes a concise list of what the module does. It also includes model assumptions and limitations
