+++ src/fswAlgorithms/attGuidance/sunSafePoint_C/_Documentation/BasiliskReportMemo.cls
@@ -118,4 +118,3 @@
 %
 %	Miscellaneous definitions
 %
-
