+++ src/fswAlgorithms/attGuidance/sunSafePoint_C/_Documentation/Figures/Basilisk-Logo.pdf
