+++ src/fswAlgorithms/attGuidance/sunSafePoint_C/_Documentation/Figures/sunHeading.pdf
