+++ src/fswAlgorithms/attGuidance/sunSafePoint_C/_Documentation/Support/sunHeading.graffle/image3.pdf
