+++ src/fswAlgorithms/attGuidance/sunSafePoint_C/_Documentation/Support/sunHeading.graffle/image5.pdf
