+++ src/fswAlgorithms/attGuidance/sunSafePoint_C/_Documentation/secUserGuide.tex
@@ -13,14 +13,13 @@ The module has 2 required input messages, and 1 output message:
 The module has the following parameter that can be configured:
 \begin{itemize}
 	\item {\tt sHatBdyCmd} -- [REQUIRED] This 3x1 array contains the commanded body-relative vector $\hat{\bm s}_{c}$ that is to be aligned with the sun heading $\bm s$
-	
-	\item {\tt minUnitMag} -- This double contains the minimum norm value of $\bm s$ such that a tracking error attitude solution $\bm\sigma_{B/R}$ is still computed.  If the norm is less than this, then $\bm\sigma_{B/R}$ is set to zero.    The default {\tt minUnitMag} value is zero. 
-	
-	\item {\tt omega\_RN\_B} -- This vectors specifies the body-fixed search rate to rotate and search for the sun if no good sun direction vector is visible.  Default value is a zero vector.  
 
-	\item {\tt smallAngle} -- This double specifies what is considered close for $\bm s$ and $\hat{\bm s}_{c}$ to be collinear.  Default value is zero.  
-	
-	\item {\tt sunAxisSpinRate} -- Specifies the nominal spin rate about the sun heading vector.  This is only used if a sun heading solution is available.  Default value is zero bring the spacecraft to rest. 
-	
+	\item {\tt minUnitMag} -- This double contains the minimum norm value of $\bm s$ such that a tracking error attitude solution $\bm\sigma_{B/R}$ is still computed.  If the norm is less than this, then $\bm\sigma_{B/R}$ is set to zero.    The default {\tt minUnitMag} value is zero.
+
+	\item {\tt omega\_RN\_B} -- This vectors specifies the body-fixed search rate to rotate and search for the sun if no good sun direction vector is visible.  Default value is a zero vector.
+
+	\item {\tt smallAngle} -- This double specifies what is considered close for $\bm s$ and $\hat{\bm s}_{c}$ to be collinear.  Default value is zero.
+
+	\item {\tt sunAxisSpinRate} -- Specifies the nominal spin rate about the sun heading vector.  This is only used if a sun heading solution is available.  Default value is zero bring the spacecraft to rest.
+
 \end{itemize}
-
