+++ src/fswAlgorithms/attGuidance/sunSafePoint_C/sunSafePoint_C.i
@@ -16,25 +16,22 @@
  OR IN CONNECTION WITH THE USE OR PERFORMANCE OF THIS SOFTWARE.
 
  */
-%module sunSafePointCpp
+%module sunSafePoint_C
 %{
-   #include "sunSafePointCpp.h"
+   #include "sunSafePoint_C.h"
 %}
 
 %pythoncode %{
-from Basilisk.architecture.swig_common_model import *
+    from Basilisk.architecture.swig_common_model import *
 %}
-%include "std_string.i"
-%include "swig_conly_data.i"
-%include "swig_eigen.i"
 
 %include "sys_model.i"
-%include "sunSafePointCpp.h"
+%include "swig_conly_data.i"
+
+%include "sunSafePoint_C.h"
 
 %include "architecture/msgPayloadDefC/NavAttMsgPayload.h"
-struct NavAttMsg_C;
 %include "architecture/msgPayloadDefC/AttGuidMsgPayload.h"
-struct AttGuidMsg_C;
 
 %pythoncode %{
 import sys
